requirements.txt
playwright==1.40.0
flask==3.0.0
flask-cors==4.0.0
pillow==10.1.0